Cloud layouter test moved and renamed
