+++ cs/TagsCloudVisualization/CircularCloudLayouter.cs
@@ -1,10 +1,11 @@
 ﻿using System.Drawing;
+using TagsCloudVisualization.Interfaces;
 
 namespace TagsCloudVisualization;
 
-public class CircularCloudLayouter(Point center) : ICircularCloudLayouter
+public class CircularCloudLayouter(Point center) : ICloudLayouter
 {
-    private readonly PointGenerator _pointGenerator = new(center);
+    private readonly IPointGenerator _pointGenerator = new SpiralPointGenerator(center);
     private readonly List<Rectangle> _rectangles = [];
 
     public Rectangle PutNextRectangle(Size size)
@@ -14,11 +15,10 @@ public class CircularCloudLayouter(Point center) : ICircularCloudLayouter
                 $"{nameof(size.Width)} and {nameof(size.Height)} should be greater than zero");
 
         Rectangle newRectangle;
-        do newRectangle = new Rectangle(_pointGenerator.GetNewPoint(), size);
+        do newRectangle = new Rectangle(_pointGenerator.GeneratePoint(), size);
         while (_rectangles.Any(rec => newRectangle.IntersectsWith(rec)));
 
         _rectangles.Add(newRectangle);
         return newRectangle;
     }
 }
-
